backend/pyproject.toml
[project]
name = "redaction-app"
version = "0.1.0"
description = "AI-powered PDF redaction tool using Gemini"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "google-genai>=0.8.0",
    "pymupdf>=1.24.0",
    "tenacity>=9.0.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "python-multipart>=0.0.22",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.11.0",
    "pyright>=1.1.0",
]

[tool.setuptools]
py-modules = ["main", "config", "retry", "redaction_agent"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["B008"]

[tool.pyright]
pythonVersion = "3.11"
pythonPlatform = "All"
venvPath = "."
venv = ".venv"
typeCheckingMode = "standard"
include = ["*.py"]
exclude = ["build", ".venv", "**/__pycache__"]
